server/requirements.txt
pandas
python-pptx
openpyxl
sentence-transformers